zephyr: include/ua_types.h: Define UA_Int32 as int as a workaround
open62541 defines UA_Int32 as int32_t, but then in many places in logging
prints it using "%i"/"%u" specifiers. But these specifiers are defined to
work with int/unsigned int, whereas int32_t/uint32_t doesn't have to be
defined as such. Various toolchains define it as long's instead. Overall,
this is bug on the part of open62541, the C standard requires use of
specical macros like PRI_I32?? to print int32_t.

With Zephyr, we exactly have a case when int32_t as long in many toolchains.
So, leaving "typedef int32_t UA_Int32" produces gazillion of warnings for
logging statements. Instead switching it to "typedef int UA_Int32" still
produces warning, but only a couple, for open62541 atomic ops functions
(which are again inconsistently defined, instead of accepting UA_UInt32 as
used everywhere else in open62541, it's defined as uint32_t, but of course
UA_UInt32 is what's actually passed).